requirement.txt
Django==2.0.5
python-decouple==3.1
mysqlclient==1.4.2.post1
pytz==2018.9
python-decouple==3.1
pillow==5.4.1
django-filter==1.1.0
django-storages==1.6.5
djangorestframework==3.7.3
